[BACKLOG-23020] - Merge master changes with karaf shim commit
