+++ angular-bundle/pom.xml
@@ -14,7 +14,6 @@
   <description>a Pentaho open source project</description>
   <url>http://www.pentaho.com</url>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
