+++ pentaho-authentication-mapper/impl/pom.xml
@@ -24,7 +24,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
