+++ pentaho-blueprint-collection-utils/pom.xml
@@ -20,7 +20,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
