+++ pentaho-i18n-bundle/example/pom.xml
@@ -26,7 +26,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
