+++ pentaho-i18n-bundle/pom.xml
@@ -97,7 +97,6 @@
         </includes>
       </resource>
     </resources>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
