+++ pentaho-i18n-webservice-bundle/pom.xml
@@ -96,7 +96,6 @@
         </includes>
       </resource>
     </resources>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
