+++ pentaho-kettle-repository-locator/api/pom.xml
@@ -30,7 +30,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
