+++ pentaho-kettle-repository-locator/impl/platform/pom.xml
@@ -53,7 +53,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
