+++ pentaho-kettle-repository-locator/impl/spoon/pom.xml
@@ -47,7 +47,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
