+++ pentaho-metastore-locator-api/src/main/java/org/pentaho/osgi/metastore/locator/api/MetastoreLocator.java
