+++ pentaho-metastore-locator-impl/local/src/main/java/org/pentaho/osgi/metastore/locator/api/impl/MetastoreLocatorNewTransExtensionPoint.java
