+++ pentaho-metastore-locator-impl/local/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<blueprint xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd"
+           xmlns:pen="http://www.pentaho.com/xml/schemas/pentaho-blueprint">
+           
+           
+    <bean id="metastoreLocatorImpl" class="org.pentaho.osgi.metastore.locator.api.impl.MetastoreLocatorImpl" scope="singleton"/>
+
+    <service ref="metastoreLocatorImpl" interface="org.pentaho.osgi.metastore.locator.api.MetastoreLocator"/>
+    <service ref="metastoreLocatorImpl" interface="org.pentaho.di.core.osgi.api.MetastoreLocatorOsgi"/>
+
+    <reference-list availability="optional" interface="org.pentaho.osgi.metastore.locator.api.MetastoreProvider">
+        <reference-listener ref="metastoreLocatorImpl" bind-method="itemAdded" unbind-method="itemRemoved"/>
+    </reference-list>
+
+    <bean id="MetastoreLocatorExtensionPoint" scope="singleton" class="org.pentaho.osgi.metastore.locator.api.impl.MetastoreLocatorExtensionPoint">
+        <argument ref="metastoreLocatorImpl" />
+        <pen:di-plugin type="org.pentaho.di.core.extension.ExtensionPointPluginType"/>
+    </bean>
+
+    <bean id="MetastoreLocatorNewTransExtensionPoint" scope="singleton" class="org.pentaho.osgi.metastore.locator.api.impl.MetastoreLocatorNewTransExtensionPoint">
+        <argument ref="metastoreLocatorImpl" />
+        <pen:di-plugin type="org.pentaho.di.core.extension.ExtensionPointPluginType"/>
+    </bean>
+
+    <bean id="MetastoreLocatorPrepareTransExtensionPoint" scope="singleton" class="org.pentaho.osgi.metastore.locator.api.impl.MetastoreLocatorPrepareTransExtensionPoint">
+        <argument ref="metastoreLocatorImpl" />
+        <pen:di-plugin type="org.pentaho.di.core.extension.ExtensionPointPluginType"/>
+    </bean>
+
+    <bean id="MetastoreLocatorBeforeJobExtensionPoint" scope="singleton" class="org.pentaho.osgi.metastore.locator.api.impl.MetastoreLocatorBeforeJobExtensionPoint">
+        <argument ref="metastoreLocatorImpl" />
+        <pen:di-plugin type="org.pentaho.di.core.extension.ExtensionPointPluginType"/>
+    </bean>
+
+    <bean id="MetastoreLocatorBeforeJobSaveExtensionPoint" scope="singleton" class="org.pentaho.osgi.metastore.locator.api.impl.MetastoreLocatorBeforeJobSaveExtensionPoint">
+        <argument ref="metastoreLocatorImpl" />
+        <pen:di-plugin type="org.pentaho.di.core.extension.ExtensionPointPluginType"/>
+    </bean>
+           
+    <bean id="localMetastoreProvider" class="org.pentaho.osgi.metastore.locator.impl.local.LocalFileMetastoreProvider" scope="singleton"/>
+    <service ref="localMetastoreProvider" interface="org.pentaho.osgi.metastore.locator.api.MetastoreProvider">
+	    <service-properties>
+	       <entry key="serviceMapKey" value="LocalMetastoreProvider"/>
+	   </service-properties>
+    </service>
+    
+</blueprint>
