+++ pentaho-metastore-locator-impl/local/src/test/java/org/pentaho/osgi/metastore/locator/api/impl/MetastoreLocatorNewTransExtensionPointTest.java
