+++ pentaho-metastore-locator-impl/local/src/test/java/org/pentaho/osgi/metastore/locator/impl/local/LocalFileMetastoreProviderTest.java
