+++ pentaho-metastore-locator-impl/repository/src/main/java/org/pentaho/osgi/metastore/locator/impl/repository/RepositoryMetastoreProvider.java
