+++ pentaho-metastore-locator-impl/repository/src/main/resources/OSGI-INF/blueprint/blueprint.xml
