+++ pentaho-notification-webservice-bundle/pom.xml
@@ -88,7 +88,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>com.github.searls</groupId>
