+++ pentaho-osgi-utils-impl/pom.xml
@@ -112,7 +112,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
