+++ pentaho-pdi-platform/pom.xml
@@ -118,7 +118,6 @@
         </includes>
       </resource>
     </resources>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
