+++ pentaho-platform-plugin-deployer/pom.xml
@@ -63,7 +63,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
