+++ pentaho-proxy-spring4/pom.xml
@@ -124,7 +124,6 @@
     </dependency>
   </dependencies>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
