+++ pentaho-webjars-deployer/pom.xml
@@ -2,7 +2,6 @@
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-
   <parent>
     <groupId>pentaho</groupId>
     <artifactId>pentaho-osgi-bundles</artifactId>
@@ -113,8 +112,6 @@
   </dependencies>
 
   <build>
-    <finalName>${project.artifactId}</finalName>
-
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
