+++ pentaho-zookeeper-fragment/pom.xml
@@ -12,7 +12,6 @@
   <version>8.2.0.0-SNAPSHOT</version>
   <packaging>bundle</packaging>
   <build>
-    <finalName>${project.artifactId}</finalName>
     <plugins>
       <plugin>
         <groupId>org.apache.felix</groupId>
