+++ pom.xml
@@ -99,7 +99,6 @@
       </plugin>
     </plugins>
   </build>
-
   <profiles>
     <profile>
       <id>lowdeps</id>
@@ -128,7 +127,8 @@
         <module>pentaho-object-tunnel</module>
         <module>pentaho-zookeeper-fragment</module>
         <module>pentaho-bundle-resource-manager</module>
-        <module>pentaho-proc-governor</module>
+        <module>pentaho-metastore-locator-api</module>
+	<module>pentaho-proc-governor</module>
       </modules>
     </profile>
     <profile>
@@ -144,7 +144,7 @@
         <module>pentaho-proxy-factory</module>
         <module>pentaho-proxy-spring4</module>
         <module>pentaho-kettle-repository-locator</module>
-        <module>pentaho-metastore-locator</module>
+        <module>pentaho-metastore-locator-impl</module>
         <module>pentaho-webcontext</module>
         <module>pentaho-pdi-platform</module>
         <module>pentaho-camel-guava-eventbus</module>
@@ -218,3 +218,4 @@
   </repositories>
 
 </project>
+
